Create README.md
EDA:

Exploratory Data Analysis (EDA) is the process of analyzing and visualizing data to gain insights about the data and discover patterns. In the case of bank churn data, EDA can be used to identify the most important factors that contribute to customer churn and how they are related to each other.

The first step in EDA is to get an overview of the data. This involves summarizing the data using descriptive statistics such as mean, median, and mode, as well as visualizing the data using charts and graphs. This will help to identify potential outliers, and understand the distribution of the data.

The next step is to look at the relationships between the different variables. This can be done by creating scatterplots, boxplots, or correlation matrices. These visualizations can help to identify potential correlations between the variables that may be influencing customer churn.

Finally, it is important to assess the relationships between the different variables and the churn rate. This can be done by creating a logistic regression model. This model can be used to identify the most important factors that are influencing customer churn and how they are related.

Logistic Regression:

Logistic regression is a type of machine learning algorithm used to predict a binary outcome (e.g. churn or no churn). The algorithm uses a set of independent variables to predict the probability of the outcome.

In the case of bank churn data, the independent variables could include customer demographics (e.g. age, income, gender), account information (e.g. account balance, number of transactions), and customer service (e.g. customer satisfaction).

The logistic regression model can then be used to identify the most important factors that are influencing customer churn and how they are related. The model can then be used to make predictions about whether a customer is likely to churn or not.

By combining EDA and logistic regression, it is possible to gain insights into the data and make more accurate predictions about customer churn.